api/requirements.txt
annotated-types==0.7.0
anyio==4.6.2.post1
bcrypt==4.2.0
certifi==2024.8.30
click==8.1.7
colorama==0.4.6
dnspython==2.7.0
fastapi==0.115.2
h11==0.14.0
httpcore==1.0.6
httpx==0.27.2
idna==3.10
iniconfig==2.0.0
mongomock==4.2.0.post1
packaging==24.1
passlib==1.7.4
pluggy==1.5.0
pydantic==2.9.2
pydantic-core==2.23.4
pymongo==4.10.1
pytest==8.3.3
python-dateutil==2.9.0.post0
pytz==2024.2
sentinels==1.0.0
six==1.16.0
sniffio==1.3.1
starlette==0.40.0
typing-extensions==4.12.2
uvicorn==0.32.0